Moved wait code in it specific handler for serialized object reception. refs #70
