+++ include/boost/mpi/detail/request_handlers.hpp
@@ -11,124 +11,10 @@
 #ifndef BOOST_MPI_REQUEST_HANDLERS_HPP
 #define BOOST_MPI_REQUEST_HANDLERS_HPP
 
+#include <boost/mpi/skeleton_and_content.hpp>
+
 namespace boost { namespace mpi {
 
-struct request::legacy_handler : public request::handler {
-  template<typename T> legacy_handler(communicator const& comm, int source, int tag, T& value);
-  template<typename T> legacy_handler(communicator const& comm, int source, int tag, T* value, int n);
-  template<typename T, class A> legacy_handler(communicator const& comm, int source, int tag, std::vector<T,A>& values, mpl::true_ primitive);
-  
-  status wait();
-  optional<status> test();
-  void cancel();
-  
-  bool active() const;
-  optional<MPI_Request&> trivial();
-  
-  MPI_Request& size_request() { return m_requests[0]; }
-  MPI_Request& payload_request() { return m_requests[1]; }
-  
-  boost::shared_ptr<void> data() { return m_data; }
-  void                    set_data(boost::shared_ptr<void> d) { m_data = d; }
-  
-  template<class T>  boost::shared_ptr<T>    data() { return boost::static_pointer_cast<T>(m_data); }
-  template<class T>  void                    set_data(boost::shared_ptr<T> d) { m_data = d; }
-  
-  typedef optional<status> (*handler_type)(request::legacy_handler* self, 
-                                           request_action action);
-  template<typename T>
-  static optional<status> 
-  handle_serialized_irecv(legacy_handler* self, request_action action);
-  
-  template<typename T>
-  static optional<status> 
-  handle_serialized_array_irecv(legacy_handler* self, request_action action);
-  
-  template<typename T, class A>
-  static optional<status> 
-  handle_dynamic_primitive_array_irecv(legacy_handler* self, request_action action);
-  
-  MPI_Request      m_requests[2];
-  shared_ptr<void> m_data;
-  handler_type     m_handler;
-};
-
-template<typename T>
-struct request::legacy_serialized_handler 
-  : public request::legacy_handler {
-  legacy_serialized_handler(communicator const& comm, int source, int tag, T& value)
-    : legacy_handler(comm, source, tag, value) {}
-};
-
-template<typename T>
-struct request::legacy_serialized_array_handler 
-  : public request::legacy_handler {
-  legacy_serialized_array_handler(communicator const& comm, int source, int tag, T* values, int n)
-    : legacy_handler(comm, source, tag, values, n) {}
-};
-
-template<typename T, class A>
-struct request::legacy_dynamic_primitive_array_handler 
-  : public request::legacy_handler {
-  legacy_dynamic_primitive_array_handler(communicator const& comm, int source, int tag, std::vector<T,A>& values)
-    : legacy_handler(comm, source, tag, values, mpl::true_()) {}
-};
-
-struct request::trivial_handler : public request::handler {
-  trivial_handler();
-  
-  status wait();
-  optional<status> test();
-  void cancel();
-  
-  bool active() const;
-  optional<MPI_Request&> trivial();
-  
-  MPI_Request& size_request();
-  MPI_Request& payload_request();
-  
-  boost::shared_ptr<void> data();
-  void                    set_data(boost::shared_ptr<void> d);
-  
-  MPI_Request      m_request;
-  shared_ptr<void> m_data;
-};
-
-struct request::dynamic_handler : public request::handler {
-  dynamic_handler();
-  
-  status wait();
-  optional<status> test();
-  void cancel();
-  
-  bool active() const;
-  optional<MPI_Request&> trivial();
-  
-  MPI_Request& size_request();
-  MPI_Request& payload_request();
-  
-  boost::shared_ptr<void> data();
-  void                    set_data(boost::shared_ptr<void> d);
-  
-  MPI_Request      m_requests[2];
-  shared_ptr<void> m_data;
-};
-
-template<typename T> 
-request request::make_serialized(communicator const& comm, int source, int tag, T& value) {
-  return request(new legacy_serialized_handler<T>(comm, source, tag, value));
-}
-
-template<typename T>
-request request::make_serialized_array(communicator const& comm, int source, int tag, T* values, int n) {
-  return request(new legacy_serialized_array_handler<T>(comm, source, tag, values, n));
-}
-
-template<typename T, class A>
-request request::make_dynamic_primitive_array(communicator const& comm, int source, int tag, 
-                                              std::vector<T,A>& values) {
-  return request(new legacy_dynamic_primitive_array_handler<T,A>(comm, source, tag, values));
-}
 
 namespace detail {
   /**
@@ -241,6 +127,180 @@ namespace detail {
     std::vector<T,A>& values;
   };
 
+  template<typename T>
+  struct serialized_irecv_data<const skeleton_proxy<T> >
+  {
+    serialized_irecv_data(const communicator& comm, int source, int tag, 
+                          skeleton_proxy<T> proxy)
+      : comm(comm), source(source), tag(tag), isa(comm), 
+        ia(isa.get_skeleton()), proxy(proxy) { }
+
+    void deserialize(status& stat) 
+    { 
+      isa >> proxy.object;
+      stat.m_count = 1;
+    }
+
+    communicator comm;
+    int source;
+    int tag;
+    std::size_t count;
+    packed_skeleton_iarchive isa;
+    packed_iarchive& ia;
+    skeleton_proxy<T> proxy;
+  };
+
+  template<typename T>
+  struct serialized_irecv_data<skeleton_proxy<T> >
+    : public serialized_irecv_data<const skeleton_proxy<T> >
+  {
+    typedef serialized_irecv_data<const skeleton_proxy<T> > inherited;
+
+    serialized_irecv_data(const communicator& comm, int source, int tag, 
+                          const skeleton_proxy<T>& proxy)
+      : inherited(comm, source, tag, proxy) { }
+  };
+}
+
+struct request::legacy_handler : public request::handler {
+  template<typename T> legacy_handler(communicator const& comm, int source, int tag, T& value);
+  template<typename T> legacy_handler(communicator const& comm, int source, int tag, T* value, int n);
+  template<typename T, class A> legacy_handler(communicator const& comm, int source, int tag, std::vector<T,A>& values, mpl::true_ primitive);
+  
+  status wait();
+  optional<status> test();
+  void cancel();
+  
+  bool active() const;
+  optional<MPI_Request&> trivial();
+  
+  MPI_Request& size_request() { return m_requests[0]; }
+  MPI_Request& payload_request() { return m_requests[1]; }
+  
+  boost::shared_ptr<void> data() { return m_data; }
+  void                    set_data(boost::shared_ptr<void> d) { m_data = d; }
+  
+  template<class T>  boost::shared_ptr<T>    data() { return boost::static_pointer_cast<T>(m_data); }
+  template<class T>  void                    set_data(boost::shared_ptr<T> d) { m_data = d; }
+  
+  typedef optional<status> (*handler_type)(request::legacy_handler* self, 
+                                           request_action action);
+  template<typename T>
+  static optional<status> 
+  handle_serialized_irecv(legacy_handler* self, request_action action);
+  
+  template<typename T>
+  static optional<status> 
+  handle_serialized_array_irecv(legacy_handler* self, request_action action);
+  
+  template<typename T, class A>
+  static optional<status> 
+  handle_dynamic_primitive_array_irecv(legacy_handler* self, request_action action);
+  
+  MPI_Request      m_requests[2];
+  shared_ptr<void> m_data;
+  handler_type     m_handler;
+};
+
+template<typename T>
+struct request::legacy_serialized_handler 
+  : public request::legacy_handler {
+  legacy_serialized_handler(communicator const& comm, int source, int tag, T& value)
+    : legacy_handler(comm, source, tag, value) {}
+  status wait() {
+    typedef detail::serialized_irecv_data<T> data_t;
+    shared_ptr<data_t> data = this->data<data_t>();
+    
+    status stat;
+    if (m_requests[1] == MPI_REQUEST_NULL) {
+      // Wait for the count message to complete
+      BOOST_MPI_CHECK_RESULT(MPI_Wait,
+                             (m_requests, &stat.m_status));
+      // Resize our buffer and get ready to receive its data
+      data->ia.resize(data->count);
+      BOOST_MPI_CHECK_RESULT(MPI_Irecv,
+                             (data->ia.address(), data->ia.size(), MPI_PACKED,
+                              stat.source(), stat.tag(), 
+                              MPI_Comm(data->comm), m_requests + 1));
+    }
+
+    // Wait until we have received the entire message
+    BOOST_MPI_CHECK_RESULT(MPI_Wait,
+                           (m_requests + 1, &stat.m_status));
+
+    data->deserialize(stat);
+    return stat;    
+  }
+};
+
+template<typename T>
+struct request::legacy_serialized_array_handler 
+  : public request::legacy_handler {
+  legacy_serialized_array_handler(communicator const& comm, int source, int tag, T* values, int n)
+    : legacy_handler(comm, source, tag, values, n) {}
+};
+
+template<typename T, class A>
+struct request::legacy_dynamic_primitive_array_handler 
+  : public request::legacy_handler {
+  legacy_dynamic_primitive_array_handler(communicator const& comm, int source, int tag, std::vector<T,A>& values)
+    : legacy_handler(comm, source, tag, values, mpl::true_()) {}
+};
+
+struct request::trivial_handler : public request::handler {
+  trivial_handler();
+  
+  status wait();
+  optional<status> test();
+  void cancel();
+  
+  bool active() const;
+  optional<MPI_Request&> trivial();
+  
+  MPI_Request& size_request();
+  MPI_Request& payload_request();
+  
+  boost::shared_ptr<void> data();
+  void                    set_data(boost::shared_ptr<void> d);
+  
+  MPI_Request      m_request;
+  shared_ptr<void> m_data;
+};
+
+struct request::dynamic_handler : public request::handler {
+  dynamic_handler();
+  
+  status wait();
+  optional<status> test();
+  void cancel();
+  
+  bool active() const;
+  optional<MPI_Request&> trivial();
+  
+  MPI_Request& size_request();
+  MPI_Request& payload_request();
+  
+  boost::shared_ptr<void> data();
+  void                    set_data(boost::shared_ptr<void> d);
+  
+  MPI_Request      m_requests[2];
+  shared_ptr<void> m_data;
+};
+
+template<typename T> 
+request request::make_serialized(communicator const& comm, int source, int tag, T& value) {
+  return request(new legacy_serialized_handler<T>(comm, source, tag, value));
+}
+
+template<typename T>
+request request::make_serialized_array(communicator const& comm, int source, int tag, T* values, int n) {
+  return request(new legacy_serialized_array_handler<T>(comm, source, tag, values, n));
+}
+
+template<typename T, class A>
+request request::make_dynamic_primitive_array(communicator const& comm, int source, int tag, 
+                                              std::vector<T,A>& values) {
+  return request(new legacy_dynamic_primitive_array_handler<T,A>(comm, source, tag, values));
 }
 
 template<typename T>
@@ -251,25 +311,7 @@ request::legacy_handler::handle_serialized_irecv(legacy_handler* self, request_a
   shared_ptr<data_t> data = static_pointer_cast<data_t>(self->m_data);
 
   if (action == ra_wait) {
-    status stat;
-    if (self->m_requests[1] == MPI_REQUEST_NULL) {
-      // Wait for the count message to complete
-      BOOST_MPI_CHECK_RESULT(MPI_Wait,
-                             (self->m_requests, &stat.m_status));
-      // Resize our buffer and get ready to receive its data
-      data->ia.resize(data->count);
-      BOOST_MPI_CHECK_RESULT(MPI_Irecv,
-                             (data->ia.address(), data->ia.size(), MPI_PACKED,
-                              stat.source(), stat.tag(), 
-                              MPI_Comm(data->comm), self->m_requests + 1));
-    }
-
-    // Wait until we have received the entire message
-    BOOST_MPI_CHECK_RESULT(MPI_Wait,
-                           (self->m_requests + 1, &stat.m_status));
-
-    data->deserialize(stat);
-    return stat;
+    std::abort();
   } else if (action == ra_test) {
     status stat;
     int flag = 0;
